Split common utility functions out to their own file
