+++ lib/sirp.rb
@@ -2,6 +2,7 @@ require 'openssl'
 require 'digest'
 require 'rbnacl/libsodium'
 require 'rbnacl'
+require 'sirp/util'
 require 'sirp/sirp'
 require 'sirp/parameters'
 require 'sirp/client'
