+++ spec/sirp_spec.rb
@@ -6,6 +6,7 @@ require 'spec_helper'
 #
 describe SIRP do
   include SIRP
+
   before :all do
     @N = '115b8b692e0e045692cf280b436735c77a5a9e8a9e7ed56c965f87db5b2a2ece3'.to_i(16)
     @g = 2
@@ -14,50 +15,6 @@ describe SIRP do
     @salt = '01ebb2496e4e8d32e6f7967ee9fec64e'
     @a = '7ec87196e320a2f8dfe8979b1992e0d34439d24471b62c40564bb4302866e1c2'.to_i(16)
     @b = '8143e2f299852a05717427ea9d87c6146e747d0da6e95f4390264e55a43ae96'.to_i(16)
-  end
-
-  context 'hex_to_bytes' do
-    it 'should calculate expected results' do
-      expect(hex_to_bytes('abcdef0123456789'))
-        .to eq [171, 205, 239, 1, 35, 69, 103, 137]
-    end
-  end
-
-  context 'num_to_hex' do
-    it 'should calculate expected results' do
-      num = 999_999_999_999
-      expect(num_to_hex(num))
-        .to eq 'e8d4a50fff'
-      expect('e8d4a50fff'.hex).to eq num
-    end
-  end
-
-  context 'sha_hex' do
-    it 'should calculate expected results for SHA1' do
-      str = 'foo'
-      str_unpacked = str.unpack('H*')[0]
-      str_sha = Digest::SHA1.hexdigest(str)
-      expect(sha_hex(str_unpacked, Digest::SHA1)).to eq str_sha
-    end
-
-    it 'should calculate expected results for SHA256' do
-      str = 'foo'
-      str_unpacked = str.unpack('H*')[0]
-      str_sha = Digest::SHA256.hexdigest(str)
-      expect(sha_hex(str_unpacked, Digest::SHA256)).to eq str_sha
-    end
-  end
-
-  context 'sha_str' do
-    it 'should calculate expected results for SHA1' do
-      expect(sha_str('foo', Digest::SHA1))
-        .to eq '0beec7b5ea3f0fdbc95d0dd47f3c5bc275da8a33'
-    end
-
-    it 'should calculate expected results for SHA256' do
-      expect(sha_str('foo', Digest::SHA256))
-        .to eq '2c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7ae'
-    end
   end
 
   context 'mod_exp' do
